pyproject.toml
[tool.poetry]
name = "peter"
version = "0.1.0"
description = ""
authors = ["Youngwook Kim <youngwook.kim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.6"
discord-py-interactions = "^4"
langchain = {extras = ["llms"], version = "^0.0.223"}
chromadb = "^0.3.26"
gitpython = "^3.1.31"
pinecone-client = "^2.2.2"
tiktoken = "^0.4.0"
pypdf = "^3.12.1"
google-search-results = "^2.4.2"
pydantic = "^1.10.11"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.4.1"
pyright = "^1.1.316"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
